Update card locking to remove approve icon
If card is locked for the stage, will change the approve icon to a check mark showing that the card is done and is in the next stage. This prevents the card from being pushed to the next stage. Remove ability to move card backwards from the toDo column in a stage.